Update script.js
Suggestion:
Not sure original lines 55 and 56 are needed (given xOffset and yOffset are already re-assigned during the drag function).   When trying to follow the solution, the redundancy is a little confusing , so wanted to suggest perhaps considering removing those reassignments of initialX and initialY on original lines 55-56 since they get overwritten anyhow in dragStart function (lines 18 and 19).